requirements.txt
python_pptx==0.6.18
setuptools==42.0.2
wikipedia==1.4.0
python_docx==0.8.10
google_images_download==2.8.0
docx==0.2.4
